+++ src/main/kotlin/codegen/FlutterXPluginInfo.kt
@@ -1,11 +1,11 @@
 package codegen
 // 自动生成的插件信息类,不要修改这个文件,否则会导致插件功能失效
 object FlutterXPluginInfo {
-    const val VERSION: String = "5.6.2.243"
+    const val VERSION: String = "5.7.0.251"
     const val CHANGELOG: String = """
-<h2>5.6.2 - 2025-05-08</h2>
+<h2>5.7.0 - 2025-05-26</h2>
 
-<ul><li>Add <code>dio</code> restart button (settings)</li><li>Optimize <code>l10n</code> editing panel, add preview text option, preview text in tree (l10n panel)</li><li>Add &quot;delete l10n key&quot; action, in right popup menu (l10n panel)</li><li>Add &quot;Rename l10n key&quot; action, in right popup menu (l10n panel)</li><li><a href="https://github.com/mdddj/dd_flutter_idea_plugin/blob/master/CHANGELOG.md">full changelog</a></li></ul>
+<ul><li>Add SP panel right menu actions (SP) <a href="https://flutterx.itbug.shop/shared-preferences.html#bo0451_6">see document</a></li><li><a href="https://github.com/mdddj/dd_flutter_idea_plugin/blob/master/CHANGELOG.md">full changelog</a></li></ul>
 """
 
 }