+++ src/main/kotlin/shop/itbug/fluttercheckversionx/window/l10n/L10nWindow.kt
@@ -25,14 +25,11 @@ import com.intellij.util.ui.components.BorderLayoutPanel
 import com.intellij.util.ui.tree.TreeUtil
 import shop.itbug.fluttercheckversionx.actions.context.HelpContextAction
 import shop.itbug.fluttercheckversionx.actions.context.SiteDocument
-import shop.itbug.fluttercheckversionx.actions.tool.FlutterL10nRunGenAction
-import shop.itbug.fluttercheckversionx.actions.tool.FlutterL10nSettingChangeAction
-import shop.itbug.fluttercheckversionx.actions.tool.FlutterL10nWindowTreeRefreshAction
+import shop.itbug.fluttercheckversionx.common.scroll
 import shop.itbug.fluttercheckversionx.config.PluginConfig
 import shop.itbug.fluttercheckversionx.i18n.PluginBundle
 import shop.itbug.fluttercheckversionx.services.*
 import shop.itbug.fluttercheckversionx.tools.emptyBorder
-import shop.itbug.fluttercheckversionx.widget.WidgetUtil
 import java.awt.Dimension
 import javax.swing.JPanel
 import javax.swing.JTree
@@ -46,18 +43,32 @@ import javax.swing.tree.DefaultTreeModel
 /**
  * l10n多语言窗口
  */
-class L10nWindow(val project: Project, val toolWindow: ToolWindow) : BorderLayoutPanel(), Disposable,
+class L10nWindow(val project: Project, val toolWindow: ToolWindow) : OnePixelSplitter(), Disposable,
     FlutterL10nService.OnL10nKeysChangedListener, TreeSelectionListener, FlutterL10nService.OnArbFileChangedListener,
     UiDataProvider {
+    val toolbarActionGroup =
+        ActionManager.getInstance().getAction("FlutterL10nKeysToolbarActionGroup") as DefaultActionGroup
+    val treeToolbar =
+        ActionManager.getInstance().createActionToolbar("FlutterL10nWindowTreeToolbar", toolbarActionGroup, false)
+            .apply {
+                component.border = JBUI.Borders.emptyRight(1)
+            }
     private val panel = MyPanel()
     private val editorContainer = JBScrollPane(panel)
     private val myTree = MyL10nKeysTree(project)
     private val service = FlutterL10nService.getInstance(project)
-    private val sp = JBSplitter().apply {
-        firstComponent = myTree.actionToolbar(project).apply {
-            border = JBUI.Borders.customLine(JBColor.border(), 0, 0, 0, 1)
+    private val dartStringTree = DartStringKeysTree(project)
+    val treePanel = object : BorderLayoutPanel() {
+        init {
+            addToLeft(treeToolbar.component)
+            addToCenter(myTree.scroll())
         }
-        secondComponent = editorContainer
+    }
+    private val sp = OnePixelSplitter().apply {
+        splitterProportionKey = "FlutterL10nWindowSplitterProportionRightKey"
+        firstComponent = editorContainer
+        secondComponent = dartStringTree.scroll()
+
     }
 
 
@@ -66,11 +77,15 @@ class L10nWindow(val project: Project, val toolWindow: ToolWindow) : BorderLayou
         project.messageBus.connect(this).subscribe(FlutterL10nService.ArbFileChanged, this)
         myTree.addTreeSelectionListener(this)
         editorContainer.border = emptyBorder()
-        addToCenter(sp)
         SwingUtilities.invokeLater {
             initTreeModel()
         }
         Disposer.register(this, myTree)
+        treeToolbar.targetComponent = toolWindow.component
+        putUserData(HelpContextAction.DataKey, SiteDocument.L10n)
+        this.firstComponent = treePanel
+        this.secondComponent = sp
+        this.splitterProportionKey = "FlutterL10nWindowSplitterProportionKey"
     }
 
 
@@ -152,7 +167,11 @@ class L10nWindow(val project: Project, val toolWindow: ToolWindow) : BorderLayou
 
     }
 
-    class MyPanel() : JBPanel<MyPanel>(VerticalLayout(12))
+    class MyPanel() : JBPanel<MyPanel>(VerticalLayout(12)) {
+        init {
+            border = JBUI.Borders.empty(12)
+        }
+    }
 }
 
 
@@ -180,6 +199,7 @@ class MyL10nKeysTree(val project: Project) : DnDAwareTree(DefaultMutableTreeNode
         this.cellRenderer = Render()
         project.messageBus.connect(this).subscribe(FlutterL10nService.ListenKeysChanged, this)
     }
+
 
     fun selectValue(): String? {
         val last = lastSelectedPathComponent as? DefaultMutableTreeNode ?: return null
@@ -276,20 +296,22 @@ class FlutterL10nKeyEditPanel(
 }
 
 
-private fun MyL10nKeysTree.actionToolbar(project: Project): JPanel {
-    val panel = ToolbarDecorator.createDecorator(this).setAddAction {
-        WidgetUtil.configTextFieldModal(
-            project, PluginBundle.get("l10n.addDialog.labelText"), "It inserts this key into all ARB files"
-        ) {
-            FlutterL10nService.getInstance(project).insetNewKey(it)
-        }
-    }.addExtraAction(FlutterL10nWindowTreeRefreshAction.getAction())
-        .addExtraAction(FlutterL10nSettingChangeAction.getInstance())
-        .addExtraAction(FlutterL10nRunGenAction.getInstance()).addExtraAction(HelpContextAction.ACTION)
-        .setPanelBorder(emptyBorder()).setToolbarBorder(emptyBorder()).setScrollPaneBorder(emptyBorder()).createPanel()
-    panel.border = emptyBorder()
-    putUserData(HelpContextAction.DataKey, SiteDocument.L10n)
-    return panel
-}
+//private fun MyL10nKeysTree.actionToolbar(project: Project): JPanel {
+//    val panel = ToolbarDecorator.createDecorator(this).setAddAction {
+//        WidgetUtil.configTextFieldModal(
+//            project, PluginBundle.get("l10n.addDialog.labelText"), "It inserts this key into all ARB files"
+//        ) {
+//            FlutterL10nService.getInstance(project).insetNewKey(it)
+//        }
+//    }.addExtraAction(FlutterL10nWindowTreeRefreshAction.getAction())
+//        .addExtraAction(FlutterL10nSettingChangeAction.getInstance())
+//        .addExtraAction(FlutterL10nRunGenAction.getInstance()).addExtraAction(HelpContextAction.ACTION)
+//        .setPanelBorder(emptyBorder()).setToolbarBorder(emptyBorder()).setScrollPaneBorder(emptyBorder()).createPanel()
+//    panel.border = emptyBorder()
+//    scrollsOnExpand = true
+//    panel.preferredSize = Dimension(200, -1)
+//    putUserData(HelpContextAction.DataKey, SiteDocument.L10n)
+//    return panel
+//}